Finally got a central mechanism for outputting text of different semantics
The output of different system parts (the core beliefstate system, the plugin loading system, the individual helper classes) are now using the capabilities of a central `UtilityBase` class, which allows for outputting formatted, colored text. Also, the output messages generated this way are always marked with the name of the emitting entity, making backtracking of problem origins easier.