Greatly enhanced prediction performance by pre-computation when loading model
The nodes/failures mappings (that are pretty much static throughout an experiment run) were calculated every time a prediction was triggered. With large trees, this can take up to several minutes. This is done in one step now when a new model is being loaded and is saved in a map for all future predictions, reducing the prediction time down to at most half a second.