Added a great deal of failure handling details to the symoblic log plugin
Failure handling (and rethrowing, in particular) was making serious problems during logging. This should, however, now be solved. Problems arose when failure handling nodes that previously were able to handle a failure tried to hand up the failure to a higher instance. The emitter/catcher mapping was then totally messed up, as the respective information was not updated accordingly.