Two enhancements: Don't annotate unsupported parameters, and note their types
Custom parameter annotations in event individuals have certain types that are supported. These consist of strings and numbers at the moment. Unsupported types would up to now result in an empty annotation. This is fixed now, a warning is displayed, but no empty annotation is created in the resulting .owl file.

Second, an `AnnotationInformation' individual is created now. All parameter types annotated throughout the whole experiment are denoted here, allowing easy comprehension of what to pay attention to when processing custom parameters.