Added throw/catch failure counter for OWL export
When the counter is != 0 after finishing an export, the number of failure throws didn't match the number of catches. This is a serious problem in semantics and will break the resulting output, so a warning in displayed in that case. If the warning isn't shown, everything is quietly assumed to be alright.