Big fix: Adjusted idle time values, solved inconsistencies
The idle time values were holding the system up where it was supposed to handle messages very quickly. On a conceptual level, this should not be a problem, but apparently ROS service requests get mixed up in their ordering when many of them are waiting for a ROS service to become available.

This effect was messing up log files, resulting in broken trees and unusable data. Putting the values down to reasonably small values (still > 0) solves that issue AND keeps CPU load below 100% on all cores. Formerly, all CPU cores would go to 100% load. Now, this only happens for one of them while the logging system is under full steam.