Added information on designators: Creation time and index designators
`Index designators' are designators that are the head of a chain of created designators, and include at least one successor in the chain. Also, they are not a successor of another designator. All designators now have the `creationTime' field, which denotes the time they were issued for the first time.