Added config option to the `symboliclog' plugin for floating point timepoints
The `symboliclog' plugin now features an individual configuration option `time-precision' that controls with which floating point precision time strings are generated. If the value is left out or set to zero, no floating point precision is generated (i.e. timepoints are integers). If a number is given, that many decimals are used for precision.