Big change in naming: `beliefstate` is now `semrec`. Please read the details.
The package `beliefstate` was renamed to `semrec`. This does not only change the directory and package name, but also internal namespaces (watch out, plugin developers!), plugin file prefixes (`libbs_plugin_...` -> `libsr_plugin_...`), some filenames (`Beliefstate.cpp` -> `SemanticHierarchyRecorder.cpp`, ...), and helper directory names (`~/bs_experimental_data` -> `~/sr_experimental_data`).

And here comes the reason:
As it turned out over the past periods of develment of the package, it more and more evolved into an actual recording tool than a full-fledged beliefstate. The original intentions also included some beliefstate management, but this functionality was taken over in large parts by the CRAM high level plan system.

The downside is, though, that this might break some people's setups when switching. This should be fixable pretty easily, as only one or two external directories changed, and the package/executable combination of the recorder is different from the beliefstate one.

So, to start the new version of the package, just run:
`rosrun semrec semrec`

If any problems come up, don't hesitate to contact me. It might be that I oversaw something when refactoring a huge amount of files. I will update the `beliefstate_client` package as well ASAP, as it depends on the naming convention of the (now) `semrec` component.